Работа с контроллерами (создание тестов)

+++ DistanceEducation/DistanceEducation/Models/Result.cs
@@ -3,12 +3,12 @@
     public class Result
     {
         public int Id { get; set; }
-        public int Itog { get; set; }
+        public double Itog { get; set; }
 
         //Вторичный ключ теста
         public int TestId { get; set; }
         public Test test { get; set; }
-       
+
         //Вторичный ключ студента
         public int StudentId { get; set; }
         public Student student { get; set; }
